refactor(predicates): extract controller predicates package
