+++ internal/predicates/predicates.go
